\name{DOTemp}
\alias{DOTemp}
\docType{data}
\title{Rhode River Street Pier--Maryland data set}
\description{Test data set from Stephens and Jennings SWProd calculator (USGS).  The data has been interpolated to make it have readings every fifteen minutes.  
  
}
\usage{data(DOTemp)}
\format{
 \describe{
 \item{DateTime}{DateTime}
 \item{Temp}{Temperature in Celcius}
 \item{DO}{Dissolved Oxygen}
 }
}
\details{
  This is only for example and should be used judiciously for any kind of ecosystem interpretation (requires zoo and chron packages).    
}
\source{
  Stephens, D.W., and Jennings, M.E., 1976, Determination of primaryproductivity and community metabolism in streams and lakes using dieloxygen measurements: U.S. Geological Survey Computer Contribution, 100p.
}
\keyword{datasets}
