\name{fmt.chron}
\alias{fmt.chron}
\title{ Format Dates }
\description{
  Used in the FUN argument of read.zoo for dates in the format mm/dd/yyyy hh:mm:ss
}
\usage{
fmt.chron(x)
}
\arguments{
  \item{x}{ Data Time Column }
}
\details{
  used internally in read.production
}
\author{ Stephen A Sefick Jr }
\seealso{ \code{\link{read.production}} }
\keyword{ chron }

