\name{lps.cfs}
\alias{lps.cfs}
\title{ Liters Per Second to Cubic Feet per Second }
\description{
  Convience Function for converting from liters per second to cubic feet per second
}
\usage{
lps.cfs(x)
}
\arguments{
  \item{x}{ Discharge in Liters per Second }
}
\value{
  Discharge in Cubic feet per second
}
\author{ Stephen A Sefick Jr }

\examples{
lps.cfs(134000)
}
\keyword{ math }

