\name{lps.cms}
\alias{lps.cms}
\title{ Liters per second to cubic meters per second }
\description{
  Conversion Function
}
\usage{
lps.cms(x)
}
\arguments{
  \item{x}{ discharge in Liters per second }
}
\details{
  single value or if zoo series - zoo object
}
\value{
  single value or if zoo series - zoo object
}
\author{ Stephen A Sefick Jr. }
\examples{
lps.cms(134000)
}
\keyword{ math }

