\name{SteamMetabolism-package}
\alias{StreamMetabolism-package}
\alias{StreamMetabolism}
\docType{package}
\title{
Functions for Calculating Stream Metabolism
}
\description{
Calculate all of the pieces of stream metabolism and rolls it all into one function for convience
}
\details{
\tabular{ll}{
Package: \tab Ecosystem.Production\cr
Type: \tab Package\cr
Version: \tab 0.01\cr
Date: \tab 2008-07-30\cr
License: \tab GPL>=2\cr
LazyLoad: \tab yes\cr
}
Calculate Stream Ecosystem Metabolism
}
\author{
Stephen A Sefick Jr <ssefick@gmail.com>

}
\references{

Odum, H. T. (1956). "Primary production in flowing waters." Limnology and Oceanography, 1: 
102-117.

Bott, T.L. 1996. Primary Productivity and community respiration. In F. R. Hauer and G. A. Lamberti (ed.), Methods in stream ecology. Academic Press, Inc., New York, N.Y.

APHA (2005). Standard Methods for the Examination of Water and Wastewater. 21st Edition. Eds: 
Eaton, A.D., L.S. Clesceri, E.W. Rice and A. E. Greenberg. Washington D.C., published 
jointly by the American Public Health Association, the American Water Works Association 
and the Water Environment Federation. 

Thyssen, N., Erlandsen, M., Jeppesen, E., Holm T. F.,1983. Modelling the reaeration capacity of low-land streams

O'Connor, D. J., and W. E. Dobbins (1958). Mechanisms of reaeration in natural  streams. Transactions of American Society of Civil Engineers, 123: 641-666.

Stephens, D.W., and Jennings, M.E., 1976, Determination of primaryproductivity and community metabolism in streams and lakes using dieloxygen measurements: U.S. Geological Survey Computer Contribution, 100p.

M.R. Grace and S.J. Imberger. 2006. "Stream Metabolism: Performing & Interpreting 
Measurements". Water Studies Centre Monash University, Murray Darling Basin Commission  and 
New South Wales Department of Environment and Climate Change.  204 pp. Accessed at 
http://www.sci.monash.edu.au/wsc/docs/tech-manual-v3.pdf 


}
\seealso{ \code{\link{cfs.lps}}, \code{\link{Cs}}, \code{\link{dC.dt}}, \code{\link{EcosystemProduction}}, \code{\link{Kt}}, \code{\link{lps.cfs}}, \code{\link{lps.cms}}, \code{\link{ODobbins}}, \code{\link{rear.flux}}, \code{\link{window.chron}}, \code{\link{DOTemp}}, \code{\link{fmt.chron}}, \code{\link{read.production}},
\code{\link{sunrise.set}},
\code{\link{EcosystemProduction.20}}}

\keyword{ package }
\examples{
##see help pages for all of the other functions
}